Move queryarticle and its builder to article pkg
